services/api/requirements.txt
boto3==1.34.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
aws-lambda-powertools==3.0.0
pydantic==2.5.0
